chore: associate bank conneciton to user and refresh automatically every two hours

+++ src/main/java/app/bpartners/api/repository/bridge/BridgeApi.java
@@ -329,7 +329,7 @@ public class BridgeApi {
     }
   }
 
-  public BridgeBank findBankById(Integer id) {
+  public BridgeBank findBankById(Long id) {
     try {
       HttpRequest request = HttpRequest.newBuilder()
           .uri(new URI(conf.getBankUrl() + "/" + id))
